coordinates converter and working map

+++ app/src/main/java/es/uv/and/vas/valenbisi/Geometry.java
@@ -2,9 +2,9 @@ package es.uv.and.vas.valenbisi;
 
 class Geometry {
     public String type = "";
-    public float[] coordinates = new float[2];
+    public double[] coordinates = new double[2];
 
-    Geometry(String type, float[] coordinates){
+    Geometry(String type, double[] coordinates){
         this.type = type;
         this.coordinates = coordinates;
     }
